Sequential display going with good performance at a world size of 600, with 1000 pink dots and 200 black dots. Pink dots are moving around at random, black dots are stationary. They don't detect each other at all.
The rand-angle function always starts everyone off moving outward in a perfect
circle, but after one day passes, everyone gets a new angle and it looks pretty
random.

Collisions with the end of the world are handled by sending the dots back to
the center. This can be the home spot when we have actual magical girls.

Going to try and make them detect each other since I think that will be
sticky, and I'm afraid if I put too much large-scale architecture in place,
I won't be able to work the more complex behavior (finding each other,
fighting, etc.) into it and I'll just have to throw things out.