Does basic display but only when started with lein run.
Had problems with add-vars-to-ns.

Original version:

Fails with Don't know how to create ISeq from: clojure.lang.Symbol
clojure.lang.RT.seqFrom (RT.java:505) when I pass in a symbol,
probably because it evaluates binding-map to a symbol (config,
e.g.) which can't be mapped over because it's a symbol. Fails with
Attempting to call unbound fn: clojure.core/unquote if I try to
evaluate binding-map with ~. Dunno why. Fine if I pass in a literal map.

Revised:

I backquoted the map form, so it's now ~@`(map ... ~binding-map). It was
doing unbound fn because the ~@ form was actually not inside the quoted
outer form since it was being spliced in, so I had to also backquote
that. That let me evaluated binding-map with ~. But now it doesn't work
with a literal map. Because when it evaluates the literal map, it tries
to evaluate the symbols which are the keys of that literal map, but
they haven't been bound yet, so they can't be evaluated, so it fails with
Unable to resolve symbol.

I don't know how to fix this other than making two separate macros.