The add-vars-to-ns macro still doesn't work.
It's not like it's vital to this project, but at this point it doesn't even
matter anymore; this is ground zero in my battle with macros. I don't
really like them (I think Clojure is a great language even if you leave
all the macro writing to library implementors), but I have to understand
them now because they task me.